+++ java-service-processor/src/test/resources/definition/TestIdInvalidPattern.java
@@ -1,0 +1,11 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+@ServiceDefinition
+interface TestIdInvalidPattern {
+
+    @ServiceId( pattern = "***" )
+    String getName();
+}