+++ java-service-processor/src/test/resources/definition/TestIdNestedClass.java
@@ -3,7 +3,7 @@ package definition;
 import nbbrd.service.ServiceDefinition;
 import nbbrd.service.ServiceId;
 
-class Ids {
+class TestIdNestedClass {
 
     @ServiceDefinition
     interface SimpleId {