+++ java-service-processor/src/test/resources/definition/TestIdNoArg.java
@@ -4,7 +4,7 @@ import nbbrd.service.ServiceDefinition;
 import nbbrd.service.ServiceId;
 
 @ServiceDefinition
-interface NoArgId {
+interface TestIdNoArg {
 
     @ServiceId
     String getName(String arg);