+++ java-service-processor/src/test/resources/definition/TestIdValidPattern.java
@@ -1,0 +1,11 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+@ServiceDefinition
+interface TestIdValidPattern {
+
+    @ServiceId( pattern = ServiceId.SCREAMING_SNAKE_CASE )
+    String getName();
+}