+++ java-service-processor/src/test/resources/definition/TestSorterUncheckedException.java
@@ -1,0 +1,13 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceSorter;
+
+import java.io.IOException;
+
+@ServiceDefinition
+interface TestSorterUncheckedException {
+
+    @ServiceSorter
+    int getCost() throws IllegalArgumentException;
+}