+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.0] - 2024-01-31
+
+### Added
+
+- Add regex pattern to ServiceId annotation [#253](https://github.com/nbbrd/java-service-util/issues/253)
+
+### Fixed
+
+- Fix missing checks on annotations ServiceId, ServiceFilter and ServiceSorter [#245](https://github.com/nbbrd/java-service-util/issues/245)
+
 ## [1.7.0] - 2023-06-26
 
 ### Added
@@ -141,7 +151,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/nbbrd/java-service-util/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/nbbrd/java-service-util/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/nbbrd/java-service-util/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/nbbrd/java-service-util/compare/v1.5.3...v1.6.0